Deleted setting id in constructor. Added additional constructors. id type is replaced with boxing type.

+++ src/main/java/model/Subject.java
@@ -5,7 +5,7 @@ public class Subject extends BaseEntity {
         super(name);
     }
 
-    public Subject(int id, String name) {
+    public Subject(Long id, String name) {
         super(id, name);
     }
 }
